requirements.txt
torch==2.7.1
torchvision==0.22.1
timm==1.0.26
numpy==1.26.4
pandas==2.3.0
scikit-learn==1.7.0
Pillow==11.2.1
matplotlib==3.10.3
seaborn==0.13.2
PyYAML==6.0.2
tqdm==4.67.1
jupyterlab==4.4.4
ipywidgets==8.1.8
